Support different manifest.json for Chrome/Firefox
Some values in the manifest.json differ between
Chrome and Firefox. Update the webpack config to
build separately for both browsers. Currently, the
builds are the same except for minor differences
between the manifest.json files. It's a bit overkill
but this makes it very easy to zip up the extension
for either browser.